+++ internal/modules/curriculum/application/usecases/approve_curriculum_usecase_test.go
@@ -60,7 +60,7 @@ func TestApproveCurriculumUseCase_HappyPath(t *testing.T) {
 	require.Len(t, audit.events, 1)
 	ev := audit.events[0]
 	assert.Equal(t, "curriculum.approved", ev.Action)
-	assert.Equal(t, int64(admin), ev.Fields["actor_user_id"])
+	assert.Equal(t, admin, ev.Fields["actor_user_id"])
 	assert.Equal(t, int64(7), ev.Fields["curriculum_id"])
 }
